requirements.txt
grpcio==1.59.2
grpcio-tools==1.59.2
protobuf==4.25.0
